requirements.txt
fastmcp
requests
pydantic
aiofiles
fastapi
uvicorn[standard]
python-dotenv
